+++ targets/frontend/src/components/references/index.ts
@@ -1,0 +1,2 @@
+export * from "./ReferencesTabs";
+export * from "./ContributionsReferencesView";