+++ targets/frontend/src/modules/agreements/api/query.ts
@@ -1,0 +1,28 @@
+import { gql } from "@urql/core";
+import { Agreement } from "../type";
+
+export const agreementQuery = gql`
+  query MyQuery($id: bpchar!) {
+    agreement: agreement_agreements_by_pk(id: $id) {
+      id
+      isSupported
+      kali_id
+      legifranceUrl
+      name
+      rootText
+      shortName
+      synonyms
+      updatedAt
+      workerNumber
+      publicationDate
+    }
+  }
+`;
+
+export type AgreementRequest = {
+  id: string;
+};
+
+export type AgreementResponse = {
+  agreement: Agreement;
+};