+++ targets/frontend/src/modules/documents/api/documents.controller.ts
@@ -6,10 +6,16 @@ import { DocumentsRepository, DocumentsService } from ".";
 import { InformationsRepository } from "../../informations";
 import { ContributionRepository } from "src/modules/contribution";
 import { ModelRepository } from "../../models/api";
+import { AgreementRepository } from "../../agreements/api";
 
 const inputSchema = z.object({
-  id: z.string().uuid(),
-  source: z.string(),
+  id: z.string(),
+  source: z.enum([
+    "information",
+    "modeles_de_courriers",
+    "contributions",
+    "conventions_collectives",
+  ]),
 });
 
 const actionSchema = z.object({
@@ -42,7 +48,8 @@ export class DocumentsController {
           new InformationsRepository(client),
           new DocumentsRepository(client),
           new ContributionRepository(client),
-          new ModelRepository(client)
+          new ModelRepository(client),
+          new AgreementRepository(client)
         );
         const cdtnId = await service.publish(
           inputs.input.id,