+++ targets/hasura/migrations/default/1706543495439_alter_table_public_agreements_add_column_is_supported/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "public"."agreements" DROP COLUMN "is_supported";