+++ targets/hasura/migrations/default/1706545115195_alter_table_public_agreements_add_column_short_name/down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "public"."agreements" DROP COLUMN "short_name";