+++ targets/hasura/migrations/default/1706624800030_alter_table_agreement_agreements_add_column_publication_date/up.sql
@@ -1,0 +1,2 @@
+alter table "public"."agreements" add column "publication_date" date
+ null;