+++ targets/hasura/migrations/default/1706680159745_update_document_initial_id_for_agreements/up.sql
@@ -1,0 +1,3 @@
+UPDATE public.documents
+SET initial_id = LPAD((document->>'num')::text, 4, '0')
+WHERE source = 'conventions_collectives';