+++ targets/hasura/migrations/default/1706700856662_alter_table_public_export_es_status_alter_column_informations/down.sql
@@ -1,0 +1,1 @@
+alter table "public"."export_es_status" rename column "documentsCount" to "informations";